Rename from cameraserver to cscore.
This is primarily to avoid header naming conflicts with wpilib.
