+++ cscore-jni.def
