+++ cscore.def
