+++ dependencies.gradle
@@ -58,155 +58,6 @@ ext.useWpiUtil = { project ->
     }
 }
 
-ext.useNetworkTables = { project -> 
-    project.tasks.create(name: "downloadNetworkTables") {
-        description == 'Downloads the networktables maven dependency.'
-        group = 'WPILib'
-        def depFolder = "${project.buildDir}/dependencies"
-        def ntZip = file("${depFolder}/NetworkTables.zip")
-        outputs.file(ntZip)
-        def networkTables
-        
-        doFirst {
-            def ntDependency = project.dependencies.create("edu.wpi.first.wpilib.networktables.cpp:NetworkTables:+:${project.isArm ? 'arm' : 'desktop'}@zip")
-            def ntConfig = project.configurations.detachedConfiguration(ntDependency)
-            ntConfig.setTransitive(false)
-            networkTables = ntConfig.files[0].canonicalFile
-        }
-
-        doLast {
-            copy {
-                from networkTables
-                rename 'NetworkTables(.+)', 'NetworkTables.zip'
-                into depFolder
-            }
-        }
-    }
-    
-    def networkTablesUnzipLocation = "${project.buildDir}/networktables"
-
-    // Create a task that will unzip the networktables files into a temporary build directory
-    project.tasks.create(name: "unzipNetworkTable", type: Copy) {
-        description = 'Unzips the networktables maven dependency so that the include files and libraries can be used'
-        group = 'WPILib'
-        dependsOn project.tasks.downloadNetworkTables
-        from zipTree(project.tasks.downloadNetworkTables.outputs.files.singleFile)
-        into networkTablesUnzipLocation
-    }
-
-    project.ext.networkTables = networkTablesUnzipLocation
-    project.ext.networkTablesInclude = "$networkTablesUnzipLocation/include"
-
-    project.ext.addNetworkTablesStaticLibraryLinks = { compileTask, linker, targetPlatform ->
-        compileTask.dependsOn project.tasks.unzipNetworkTable
-        String path = project.getPlatformPath2(targetPlatform)
-        if (targetPlatform.operatingSystem.windows) {
-            linker.args "${project.networkTables}/${path}/networktables.lib"
-        } else {
-            linker.args "${project.networkTables}/${path}/libnetworktables.a"
-        }
-    }
-    
-    project.ext.addNetworkTablesSharedLibraryLinks = { compileTask, linker, targetPlatform ->
-        compileTask.dependsOn project.tasks.unzipNetworkTable
-        String path = project.getPlatformPath2(targetPlatform)
-        if (targetPlatform.operatingSystem.windows) {
-            linker.args "${project.networkTables}/${path}/networktables.dll"
-        } else {
-            linker.args "${project.networkTables}/${path}/libnetworktables.so"
-        }
-    }
-}
-
-ext.useWpiLib = { project -> 
-    if (project.isArm) {
-        project.tasks.create(name: "downloadWPILib") {
-            description == 'Downloads the wpilibc maven dependency.'
-            group = 'WPILib'
-            def depFolder = "${project.buildDir}/dependencies"
-            def wpilibZip = file("${depFolder}/athena-wpilib.zip")
-            outputs.file(wpilibZip)
-            def wpilib
-            
-            doFirst {
-                def wpiLibDependency = project.dependencies.create("edu.wpi.first.wpilibc:athena:2017.+@zip")
-                def wpiLibConfig = project.configurations.detachedConfiguration(wpiLibDependency)
-                wpiLibConfig.setTransitive(false)
-                wpilib = wpiLibConfig.files[0].canonicalFile
-            }
-
-            doLast {
-                copy {
-                    from wpilib
-                    rename 'athena(.+)', 'athena-wpilib.zip'
-                    into depFolder
-                }
-            }
-        }
-        
-        project.tasks.create(name: "downloadHal") {
-            description == 'Downloads the hal maven dependency.'
-            group = 'WPILib'
-            def depFolder = "${project.buildDir}/dependencies"
-            def halZip = file("${depFolder}/hal.zip")
-            outputs.file(halZip)
-            def hal
-            
-            doFirst {
-                def halDependency = project.dependencies.create("edu.wpi.first.wpilib:hal:2017.+@zip")
-                def halConfig = project.configurations.detachedConfiguration(halDependency)
-                halConfig.setTransitive(false)
-                hal = halConfig.files[0].canonicalFile
-            }
-
-            doLast {
-                copy {
-                    from hal
-                    rename 'hal(.+)', 'hal.zip'
-                    into depFolder
-                }
-            }
-        }
-        
-        def wpiLibUnzipLocation = "${project.buildDir}/wpilibc"
-        def halUnzipLocation = "${project.buildDir}/hal"
-
-        // Create a task that will unzip the wpilib files into a temporary build directory
-        project.tasks.create(name: "unzipWpiLib", type: Copy) {
-            description = 'Unzips the wpilibc maven dependency so that the include files and libraries can be used'
-            group = 'WPILib'
-            dependsOn project.tasks.downloadWPILib
-            from zipTree(project.tasks.downloadWPILib.outputs.files.singleFile)
-            into wpiLibUnzipLocation
-        }
-        
-        // Create a task that will unzip the hal files into a temporary build directory
-        project.tasks.create(name: "unzipHal", type: Copy) {
-            description = 'Unzips the hal maven dependency so that the include files and libraries can be used'
-            group = 'WPILib'
-            dependsOn project.tasks.downloadHal
-            from zipTree(project.tasks.downloadHal.outputs.files.singleFile)
-            into halUnzipLocation
-        }        
-
-        project.ext.wpiLib = wpiLibUnzipLocation
-        project.ext.wpiLibInclude = "$wpiLibUnzipLocation/include"
-        
-        project.ext.hal = halUnzipLocation
-        project.ext.halInclude = "$halUnzipLocation/include"
-
-        project.ext.addWpiLibLibraryLinks = { compileTask, linker, targetPlatform ->
-            compileTask.dependsOn project.tasks.unzipWpiLib
-            compileTask.dependsOn project.tasks.unzipHal
-            
-            linker.args "{project.hal}/libHALAthena.so"
-            linker.args "{project.hal}/libnilibraries.so"
-            linker.args "${project.wpiLib}/libwpilibc.so"
-            linker.args << '-L' + halUnzipLocation
-            linker.args << '-L' + wpiLibUnzipLocation
-        }
-    }
-}
 ext.getOpenCvPlatformPackage = { targetPlatform ->
     if (targetPlatform.architecture.arm) {
         return 'linux-arm'
