+++ examples/enum_usb/enum_usb.cpp
@@ -1,4 +1,4 @@
-#include "cameraserver.h"
+#include "cscore.h"
 
 #include "llvm/SmallString.h"
 #include "llvm/raw_ostream.h"
