+++ java/src/edu/wpi/cscore/CvSource.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.cameraserver;
+package edu.wpi.cscore;
 
 import org.opencv.core.Mat;
