+++ java/src/edu/wpi/cscore/VideoListener.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.cameraserver;
+package edu.wpi.cscore;
 
 /// An event listener.  This calls back to a desigated callback function when
 /// an event matching the specified mask is generated by the library.
