+++ java/src/edu/wpi/cscore/VideoListenerFunction.java
@@ -5,14 +5,8 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.cameraserver;
+package edu.wpi.cscore;
 
-/// A source that represents a MJPEG-over-HTTP (IP) camera.
-public class HTTPCamera extends VideoSource {
-  /// Create a source for a MJPEG-over-HTTP (IP) camera.
-  /// @param name Source name (arbitrary unique identifier)
-  /// @param url Camera URL (e.g. "http://10.x.y.11/video/stream.mjpg")
-  public HTTPCamera(String name, String url) {
-    super(CameraServerJNI.createHTTPCamera(name, url));
-  }
+public interface VideoListenerFunction {
+  void apply(VideoEvent event);
 }
