+++ java/src/edu/wpi/cscore/VideoMode.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.cameraserver;
+package edu.wpi.cscore;
 
 /// Video mode
 public class VideoMode {
