+++ java/src/edu/wpi/cscore/VideoProperty.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.cameraserver;
+package edu.wpi.cscore;
 
 public class VideoProperty {
   public enum Type {
