+++ src/USBCameraImpl.cpp
@@ -33,7 +33,7 @@
 #include "llvm/SmallString.h"
 #include "support/timestamp.h"
 
-#include "cameraserver_cpp.h"
+#include "cscore_cpp.h"
 #include "c_util.h"
 #include "Handle.h"
 #include "Log.h"
