+++ src/cscore_c.cpp
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-#include "cameraserver_c.h"
+#include "cscore_c.h"
 
 #include <cstddef>
 #include <cstdlib>
@@ -13,7 +13,7 @@
 #include "opencv2/core/core.hpp"
 #include "llvm/SmallString.h"
 
-#include "cameraserver_cpp.h"
+#include "cscore_cpp.h"
 #include "c_util.h"
 
 extern "C" {
